+++ media_player.c
@@ -90,7 +90,7 @@ cXinelibPlayer::cXinelibPlayer(const char *File, bool Queue, const char *SubFile
   m_Speed = 1;
 
   if(File) {
-    int len = strlen(File);
+    size_t len = strlen(File);
     if(len && File[len-1] == '/') { 
       // whole directory, create temporary playlist
       m_Playlist.Read(File, true);
@@ -487,7 +487,7 @@ void cXinelibPlayerControl::Queue(const char *File)
     OpenPlayer(File, true);
     cControl::Launch(new cXinelibPlayerControl(ShowMusic, NULL));
   } else {
-    int len = strlen(File);
+    size_t len = strlen(File);
     if(len && File[len-1] == '/')
       m_Player->Playlist().Read(File, true);
     else
